storage: split upsert metrics into their own file
